Mangler Search til at virke

+++ app/src/main/java/com/example/techtrackr/data/model/PriceHistory.kt
@@ -2,8 +2,8 @@ package com.example.techtrackr.data.model
 
 data class PriceHistoryResponse(
     val productId: String?,
-    val lowest: Int?,
-    val highest: Int?,
+    val lowest: Double?,
+    val highest: Double?,
     val currencyCode: String?,
     val history: List<PriceHistoryEntry>?,
     val merchants: List<Merchant>?
